requirements.txt
Package                  Version
------------------------ -----------
absl-py                  2.2.2
aiohappyeyeballs         2.6.1
aiohttp                  3.12.15
aiosignal                1.4.0
anyio                    4.9.0
argcomplete              3.6.2
asttokens                3.0.0
astunparse               1.6.3
async-timeout            5.0.1
attrs                    25.3.0
biopython                1.85
biotite                  1.2.0
biotraj                  1.2.2
boto                     2.49.0
Brotli                   1.1.0
cachetools               5.5.2
catboost                 1.2.8
causal-conv1d            1.5.0.post8
certifi                  2025.4.26
cffi                     1.17.1
charset-normalizer       3.4.2
clarabel                 0.11.1
cloudpathlib             0.21.1
cloudpickle              3.1.1
crcmod                   1.7
cryptography             43.0.3
cvxpy                    1.7.2
cycler                   0.12.1
decorator                5.2.1
einops                   0.8.1
esm                      3.2.1
exceptiongroup           1.3.0
executing                2.2.0
fasteners                0.20
filelock                 3.18.0
flatbuffers              25.2.10
frozenlist               1.7.0
fsspec                   2025.3.2
gast                     0.6.0
gcs-oauth2-boto-plugin   3.3
google-apitools          0.5.32
google-auth              2.39.0
google-auth-httplib2     0.2.0
google-pasta             0.2.0
google-reauth            0.1.1
grpcio                   1.71.0
gsutil                   5.35
gym                      0.26.2
gym-notices              0.1.0
h11                      0.16.0
h5py                     3.13.0
httpcore                 1.0.9
httplib2                 0.20.4
httpx                    0.28.1
huggingface-hub          0.31.2
HyperPyYAML              1.2.2
idna                     3.10
imbalanced-learn         0.13.0
imblearn                 0.0
ipython                  8.37.0
jedi                     0.19.2
Jinja2                   3.1.6
joblib                   1.5.0
keras                    3.9.2
kiwisolver               1.4.9
libclang                 18.1.1
lightgbm                 4.6.0
mamba-ssm                2.2.4
Markdown                 3.8
markdown-it-py           3.0.0
MarkupSafe               3.0.2
matplotlib               3.10.7
matplotlib-inline        0.1.7
mdurl                    0.1.2
ml_dtypes                0.5.1
monotonic                1.6
mpmath                   1.3.0
msgpack                  1.1.1
msgpack-numpy            0.4.8
multidict                6.6.4
namex                    0.0.9
networkx                 3.4.2
ninja                    1.11.1.4
numpy                    2.1.3
nvidia-cublas-cu12       12.4.5.8
nvidia-cuda-cupti-cu12   12.4.127
nvidia-cuda-nvrtc-cu12   12.4.127
nvidia-cuda-runtime-cu12 12.4.127
nvidia-cudnn-cu12        9.1.0.70
nvidia-cufft-cu12        11.2.1.3
nvidia-cufile-cu12       1.11.1.6
nvidia-curand-cu12       10.3.5.147
nvidia-cusolver-cu12     11.6.1.9
nvidia-cusparse-cu12     12.3.1.170
nvidia-cusparselt-cu12   0.6.2
nvidia-nccl-cu12         2.21.5
nvidia-nvjitlink-cu12    12.4.127
nvidia-nvtx-cu12         12.4.127
oauth2client             4.1.3
opt_einsum               3.4.0
optree                   0.15.0
osqp                     1.0.4
packaging                25.0
pandas                   2.2.3
parso                    0.8.4
pexpect                  4.9.0
pillow                   11.2.1
pip                      25.1
prompt_toolkit           3.0.51
propcache                0.3.2
protobuf                 5.29.4
psutil                   7.1.0
ptyprocess               0.7.0
pure_eval                0.2.3
pyasn1                   0.6.1
pyasn1_modules           0.4.2
pycparser                2.22
Pygments                 2.19.1
pyOpenSSL                24.2.1
pyparsing                3.2.3
python-dateutil          2.9.0.post0
pytorch-metric-learning  2.8.1
pytz                     2025.2
pyu2f                    0.1.5
PyYAML                   6.0.2
regex                    2024.11.6
requests                 2.32.3
retry_decorator          1.1.1
rich                     14.0.0
rsa                      4.7.2
ruamel.yaml              0.18.12
ruamel.yaml.clib         0.2.12
safetensors              0.5.3
scikit-learn             1.6.1
scipy                    1.15.3
scs                      3.2.8
seaborn                  0.13.2
sentencepiece            0.2.0
seqtk                    0.5.0
setuptools               78.1.1
six                      1.17.0
sklearn-compat           0.1.3
sniffio                  1.3.1
speechbrain              1.0.3
stack-data               0.6.3
sympy                    1.13.1
tenacity                 9.1.2
tensorboard              2.19.0
tensorboard-data-server  0.7.2
termcolor                3.1.0
threadpoolctl            3.6.0
tokenizers               0.21.1
torch                    2.6.0
torchaudio               2.6.0
torchtext                0.18.0
torchvision              0.21.0
tqdm                     4.67.1
traitlets                5.14.3
transformers             4.48.1
triton                   3.2.0
typing_extensions        4.13.2
tzdata                   2025.2
urllib3                  2.4.0
wcwidth                  0.2.13
Werkzeug                 3.1.3
wheel                    0.45.1
wrapt                    1.17.2
xgboost                  3.1.1
yarl                     1.20.1
zstd                     1.5.7.2